+++ data/src/main/kotlin/com/wepli/data/artist/ArtistApi.kt
@@ -1,7 +1,7 @@
 package com.wepli.data.artist
 
 import com.wepli.data.artist.response.ArtistListResponse
-import com.wepli.core.kotlin.FlowResult
+import com.wepli.core.kotlin.flow.FlowResult
 import retrofit2.http.GET
 
 interface ArtistApi {
