+++ data/src/main/kotlin/com/wepli/data/artist/datasource/remote/ArtistDataSource.kt
@@ -1,7 +1,7 @@
 package com.wepli.data.artist.datasource.remote
 
 import com.wepli.data.artist.response.ArtistListResponse
-import com.wepli.core.kotlin.FlowResult
+import com.wepli.core.kotlin.flow.FlowResult
 
 interface ArtistDataSource {
