+++ data/src/main/kotlin/com/wepli/data/artist/datasource/remote/ArtistDataSourceImpl.kt
@@ -2,7 +2,7 @@ package com.wepli.data.artist.datasource.remote
 
 import com.wepli.data.artist.ArtistApi
 import com.wepli.data.artist.response.ArtistListResponse
-import com.wepli.core.kotlin.FlowResult
+import com.wepli.core.kotlin.flow.FlowResult
 import javax.inject.Inject
 
 class ArtistDataSourceImpl @Inject constructor(
