+++ data/src/main/kotlin/com/wepli/data/chart/ChartApi.kt
@@ -1,7 +1,7 @@
 package com.wepli.data.chart
 
 import com.wepli.data.chart.response.ChartListResponse
-import com.wepli.core.kotlin.FlowResult
+import com.wepli.core.kotlin.flow.FlowResult
 import retrofit2.http.GET
 
 interface ChartApi {
