+++ data/src/main/kotlin/com/wepli/data/chart/datasource/remote/ChartDataSourceImpl.kt
@@ -2,7 +2,7 @@ package com.wepli.data.chart.datasource.remote
 
 import com.wepli.data.chart.ChartApi
 import com.wepli.data.chart.response.ChartListResponse
-import com.wepli.core.kotlin.FlowResult
+import com.wepli.core.kotlin.flow.FlowResult
 import javax.inject.Inject
 
 class ChartDataSourceImpl @Inject constructor(
