+++ data/src/main/kotlin/com/wepli/data/chart/repository/ChartRepositoryImpl.kt
@@ -3,7 +3,7 @@ package com.wepli.data.chart.repository
 import com.wepli.data.chart.datasource.remote.ChartDataSource
 import com.wepli.data.chart.response.toEntities
 import com.wepli.data.network.toEntityResult
-import com.wepli.core.kotlin.FlowResult
+import com.wepli.core.kotlin.flow.FlowResult
 import com.wepli.data.di.qualifier.SupabaseDataSource
 import model.music.ChartMusic
 import repository.chart.ChartRepository
