+++ data/src/main/kotlin/com/wepli/data/keyword/datasource/KeywordDatasource.kt
@@ -1,6 +1,6 @@
 package com.wepli.data.keyword.datasource
 
-import com.wepli.core.kotlin.FlowResult
+import com.wepli.core.kotlin.flow.FlowResult
 import com.wepli.data.keyword.response.RecommendKeywordResponse
 
 interface KeywordDatasource {
