+++ data/src/main/kotlin/com/wepli/data/keyword/datasource/KeywordSupabaseDatasourceImpl.kt
@@ -1,6 +1,6 @@
 package com.wepli.data.keyword.datasource
 
-import com.wepli.core.kotlin.FlowResult
+import com.wepli.core.kotlin.flow.FlowResult
 import com.wepli.data.keyword.response.RecommendKeywordResponse
 import io.github.jan.supabase.SupabaseClient
 import io.github.jan.supabase.postgrest.postgrest
