+++ data/src/main/kotlin/com/wepli/data/relaylist/RelaylistApi.kt
@@ -1,6 +1,6 @@
 package com.wepli.data.relaylist
 
-import com.wepli.core.kotlin.FlowResult
+import com.wepli.core.kotlin.flow.FlowResult
 import com.wepli.data.relaylist.response.RelaylistResponse
 import retrofit2.http.GET
