+++ data/src/main/kotlin/com/wepli/data/song/datasource/SongDataSource.kt
@@ -1,6 +1,6 @@
 package com.wepli.data.song.datasource
 
-import com.wepli.core.kotlin.FlowResult
+import com.wepli.core.kotlin.flow.FlowResult
 import com.wepli.data.common.supabase.response.IdResponse
 import com.wepli.data.song.request.SongRequestBody
