+++ data/src/main/kotlin/com/wepli/data/song/repository/SongRepositoryImpl.kt
@@ -1,6 +1,6 @@
 package com.wepli.data.song.repository
 
-import com.wepli.core.kotlin.FlowResult
+import com.wepli.core.kotlin.flow.FlowResult
 import com.wepli.data.network.toEntityResult
 import com.wepli.data.song.datasource.SongDataSource
 import com.wepli.data.song.request.SongRequestBody
