+++ domain/src/main/kotlin/model/recommend/repository/KeywordRepository.kt
@@ -1,6 +1,6 @@
 package model.recommend.repository
 
-import com.wepli.core.kotlin.FlowResult
+import com.wepli.core.kotlin.flow.FlowResult
 import model.recommend.RecommendKeyword
 
 interface KeywordRepository {
