+++ domain/src/main/kotlin/repository/artist/ArtistRepository.kt
@@ -1,6 +1,6 @@
 package repository.artist
 
-import com.wepli.core.kotlin.FlowResult
+import com.wepli.core.kotlin.flow.FlowResult
 import model.artist.Artist
 
 interface ArtistRepository {
