+++ domain/src/main/kotlin/repository/playlist/PlaylistRepository.kt
@@ -1,6 +1,6 @@
 package repository.playlist
 
-import com.wepli.core.kotlin.FlowResult
+import com.wepli.core.kotlin.flow.FlowResult
 import model.playlist.Playlist
 import model.playlist.RecommendPlaylist
