+++ domain/src/main/kotlin/repository/post/PostRepository.kt
@@ -1,6 +1,6 @@
 package repository.post
 
-import com.wepli.core.kotlin.FlowResult
+import com.wepli.core.kotlin.flow.FlowResult
 import model.community.Post
 
 interface PostRepository {
