+++ domain/src/main/kotlin/repository/relaylist/RelaylistRepository.kt
@@ -1,6 +1,6 @@
 package repository.relaylist
 
-import com.wepli.core.kotlin.FlowResult
+import com.wepli.core.kotlin.flow.FlowResult
 import model.relaylist.Relaylist
 
 interface RelaylistRepository {
