+++ domain/src/main/kotlin/repository/user/UserRepository.kt
@@ -1,6 +1,6 @@
 package repository.user
 
-import com.wepli.core.kotlin.FlowResult
+import com.wepli.core.kotlin.flow.FlowResult
 import model.user.User
 import java.time.Instant
