+++ feature/home/src/main/java/com/wepli/home/viewmodel/HomeViewModel.kt
@@ -1,7 +1,7 @@
 package com.wepli.home.viewmodel
 
 import base.BaseMviViewModel
-import com.wepli.core.kotlin.suspendCollectResult
+import com.wepli.core.kotlin.flow.suspendCollectResult
 import com.wepli.home.mvi.HomeEffect
 import com.wepli.home.mvi.HomeIntent
 import com.wepli.home.mvi.HomeUiState
